refactor(ci): split ordermatch tests and add workflow display names
- Move cross-chain best_orders tests from docker_ordermatch_tests.rs
  to docker_tests_inner.rs (now under docker-tests-integration feature)
- Remove ETH imports from docker_ordermatch_tests.rs (now UTXO-only)
- Update mod.rs feature gate for docker_tests_inner to docker-tests-integration
- Update CI ordermatch job to start only UTXO nodes (not ETH)
- Add friendly display names to all CI jobs using Category / Subcategory pattern:
  - Unit / {Linux, macOS, Windows}
  - Integration / {Linux, macOS, Windows}
  - Docker / {SLP, Sia, ETH, Ordermatch (UTXO), Swaps (UTXO), Watchers (UTXO),
              QRC20, Tendermint, ZCoin, Integration (Cross-chain)}
  - WASM
