+++ mm2src/mm2_main/tests/docker_tests/docker_tests_inner.rs
@@ -17,9 +17,12 @@ use crate::docker_tests::helpers::utxo::generate_utxo_coin_with_privkey;
 use crate::integration_tests_common::*;
 use common::block_on;
 use crypto::privkey::key_pair_from_seed;
+use mm2_number::BigDecimal;
 use mm2_test_helpers::for_tests::{
-    enable_eth_coin, erc20_dev_conf, eth_dev_conf, mm_dump, mycoin1_conf, mycoin_conf, MarketMakerIt, Mm2TestConf,
+    best_orders_v2, best_orders_v2_by_number, enable_eth_coin, erc20_dev_conf, eth_dev_conf, mm_dump, my_balance,
+    mycoin1_conf, mycoin_conf, MarketMakerIt, Mm2TestConf,
 };
+use mm2_test_helpers::structs::BestOrdersResponse;
 use mm2_test_helpers::{get_passphrase, structs::*};
 
 // =============================================================================
@@ -429,6 +432,453 @@ fn test_orderbook_depth() {
     assert!(rc.0.is_success(), "!orderbook: {}", rc.1);
 
     request_and_check_orderbook_depth(&mm_alice);
+
+    block_on(mm_bob.stop()).unwrap();
+    block_on(mm_alice.stop()).unwrap();
+}
+
+// =============================================================================
+// Cross-Chain Best Orders Tests
+// These tests verify best_orders RPC across ETH and UTXO coins
+// =============================================================================
+
+fn get_bob_alice() -> (MarketMakerIt, MarketMakerIt) {
+    let bob_priv_key = random_secp256k1_secret();
+    let alice_priv_key = random_secp256k1_secret();
+
+    generate_utxo_coin_with_privkey("MYCOIN", 1000.into(), bob_priv_key);
+    generate_utxo_coin_with_privkey("MYCOIN1", 1000.into(), bob_priv_key);
+    fill_eth_erc20_with_private_key(bob_priv_key);
+
+    generate_utxo_coin_with_privkey("MYCOIN", 1000.into(), alice_priv_key);
+    generate_utxo_coin_with_privkey("MYCOIN1", 1000.into(), alice_priv_key);
+    fill_eth_erc20_with_private_key(alice_priv_key);
+
+    let coins = json!([mycoin_conf(1000), mycoin1_conf(1000), eth_dev_conf(),]);
+
+    let bob_conf = Mm2TestConf::seednode(&format!("0x{}", hex::encode(bob_priv_key)), &coins);
+    let mm_bob = MarketMakerIt::start(bob_conf.conf, bob_conf.rpc_password, None).unwrap();
+
+    let (_bob_dump_log, _bob_dump_dashboard) = mm_bob.mm_dump();
+    log!("Bob log path: {}", mm_bob.log_path.display());
+
+    let alice_conf = Mm2TestConf::light_node(
+        &format!("0x{}", hex::encode(alice_priv_key)),
+        &coins,
+        &[&mm_bob.ip.to_string()],
+    );
+    let mm_alice = MarketMakerIt::start(alice_conf.conf, alice_conf.rpc_password, None).unwrap();
+
+    let (_alice_dump_log, _alice_dump_dashboard) = mm_alice.mm_dump();
+    log!("Alice log path: {}", mm_alice.log_path.display());
+
+    log!("{:?}", block_on(enable_native(&mm_bob, "MYCOIN", &[], None)));
+    log!("{:?}", block_on(enable_native(&mm_alice, "MYCOIN", &[], None)));
+
+    log!("{:?}", block_on(enable_native(&mm_bob, "MYCOIN1", &[], None)));
+    log!("{:?}", block_on(enable_native(&mm_alice, "MYCOIN1", &[], None)));
+
+    let swap_contract = swap_contract_checksum();
+    dbg!(block_on(enable_eth_coin(
+        &mm_bob,
+        "ETH",
+        &[GETH_RPC_URL],
+        &swap_contract,
+        None,
+        false
+    )));
+    dbg!(block_on(enable_eth_coin(
+        &mm_alice,
+        "ETH",
+        &[GETH_RPC_URL],
+        &swap_contract,
+        None,
+        false
+    )));
+
+    (mm_bob, mm_alice)
+}
+
+#[test]
+fn test_best_orders() {
+    let (mut mm_bob, mm_alice) = get_bob_alice();
+
+    // issue sell request on Bob side by setting base/rel price
+    log!("Issue bob sell requests");
+
+    let bob_orders = [
+        // (base, rel, price, volume, min_volume)
+        ("MYCOIN", "MYCOIN1", "0.9", "0.9", None),
+        ("MYCOIN", "MYCOIN1", "0.8", "0.9", None),
+        ("MYCOIN", "MYCOIN1", "0.7", "0.9", Some("0.9")),
+        ("MYCOIN", "ETH", "0.8", "0.9", None),
+        ("MYCOIN1", "MYCOIN", "0.8", "0.9", None),
+        ("MYCOIN1", "MYCOIN", "0.9", "0.9", None),
+        ("ETH", "MYCOIN", "0.8", "0.9", None),
+        ("MYCOIN1", "ETH", "0.8", "0.8", None),
+        ("MYCOIN1", "ETH", "0.7", "0.8", Some("0.8")),
+    ];
+    for (base, rel, price, volume, min_volume) in bob_orders.iter() {
+        let rc = block_on(mm_bob.rpc(&json! ({
+            "userpass": mm_bob.userpass,
+            "method": "setprice",
+            "base": base,
+            "rel": rel,
+            "price": price,
+            "volume": volume,
+            "min_volume": min_volume.unwrap_or("0.00777"),
+            "cancel_previous": false,
+        })))
+        .unwrap();
+        assert!(rc.0.is_success(), "!setprice: {}", rc.1);
+    }
+
+    block_on(mm_bob.wait_for_log(22., |log| {
+        log.contains("DEBUG Handling IncludedTorelaysMesh message for peer")
+    }))
+    .unwrap();
+
+    let rc = block_on(mm_alice.rpc(&json! ({
+        "userpass": mm_alice.userpass,
+        "method": "best_orders",
+        "coin": "MYCOIN",
+        "action": "buy",
+        "volume": "0.1",
+    })))
+    .unwrap();
+    assert!(rc.0.is_success(), "!best_orders: {}", rc.1);
+    let response: BestOrdersResponse = serde_json::from_str(&rc.1).unwrap();
+    let best_mycoin1_orders = response.result.get("MYCOIN1").unwrap();
+    assert_eq!(1, best_mycoin1_orders.len());
+    let expected_price: BigDecimal = "0.8".parse().unwrap();
+    assert_eq!(expected_price, best_mycoin1_orders[0].price);
+
+    let rc = block_on(mm_alice.rpc(&json! ({
+        "userpass": mm_alice.userpass,
+        "method": "best_orders",
+        "coin": "MYCOIN",
+        "action": "buy",
+        "volume": "1.7",
+    })))
+    .unwrap();
+    assert!(rc.0.is_success(), "!best_orders: {}", rc.1);
+    let response: BestOrdersResponse = serde_json::from_str(&rc.1).unwrap();
+    // MYCOIN1
+    let best_mycoin1_orders = response.result.get("MYCOIN1").unwrap();
+    let expected_price: BigDecimal = "0.7".parse().unwrap();
+    let bob_mycoin1_addr = block_on(my_balance(&mm_bob, "MYCOIN1")).address;
+    // let bob_mycoin1_addr = mm_bob.display_address("MYCOIN1").unwrap();
+    assert_eq!(expected_price, best_mycoin1_orders[0].price);
+    assert_eq!(bob_mycoin1_addr, best_mycoin1_orders[0].address);
+    let expected_price: BigDecimal = "0.8".parse().unwrap();
+    assert_eq!(expected_price, best_mycoin1_orders[1].price);
+    assert_eq!(bob_mycoin1_addr, best_mycoin1_orders[1].address);
+    // ETH
+    let expected_price: BigDecimal = "0.8".parse().unwrap();
+    let best_eth_orders = response.result.get("ETH").unwrap();
+    assert_eq!(expected_price, best_eth_orders[0].price);
+
+    let rc = block_on(mm_alice.rpc(&json! ({
+        "userpass": mm_alice.userpass,
+        "method": "best_orders",
+        "coin": "MYCOIN",
+        "action": "sell",
+        "volume": "0.1",
+    })))
+    .unwrap();
+    assert!(rc.0.is_success(), "!best_orders: {}", rc.1);
+    let response: BestOrdersResponse = serde_json::from_str(&rc.1).unwrap();
+
+    let expected_price: BigDecimal = "1.25".parse().unwrap();
+
+    let best_mycoin1_orders = response.result.get("MYCOIN1").unwrap();
+    assert_eq!(expected_price, best_mycoin1_orders[0].price);
+    assert_eq!(1, best_mycoin1_orders.len());
+
+    let best_eth_orders = response.result.get("ETH").unwrap();
+    assert_eq!(expected_price, best_eth_orders[0].price);
+
+    let rc = block_on(mm_alice.rpc(&json! ({
+        "userpass": mm_alice.userpass,
+        "method": "best_orders",
+        "coin": "ETH",
+        "action": "sell",
+        "volume": "0.1",
+    })))
+    .unwrap();
+    assert!(rc.0.is_success(), "!best_orders: {}", rc.1);
+    let response: BestOrdersResponse = serde_json::from_str(&rc.1).unwrap();
+
+    let expected_price: BigDecimal = "1.25".parse().unwrap();
+
+    let best_mycoin1_orders = response.result.get("MYCOIN1").unwrap();
+    assert_eq!(expected_price, best_mycoin1_orders[0].price);
+    assert_eq!("MYCOIN1", best_mycoin1_orders[0].coin);
+    assert_eq!(1, best_mycoin1_orders.len());
+
+    block_on(mm_bob.stop()).unwrap();
+    block_on(mm_alice.stop()).unwrap();
+}
+
+#[test]
+fn test_best_orders_v2_by_number() {
+    let (mut mm_bob, mm_alice) = get_bob_alice();
+
+    // issue sell request on Bob side by setting base/rel price
+    log!("Issue bob sell requests");
+
+    let bob_orders = [
+        // (base, rel, price, volume, min_volume)
+        ("MYCOIN", "MYCOIN1", "0.9", "0.9", None),
+        ("MYCOIN", "MYCOIN1", "0.8", "0.9", None),
+        ("MYCOIN", "MYCOIN1", "0.7", "0.9", Some("0.9")),
+        ("MYCOIN", "ETH", "0.8", "0.9", None),
+        ("MYCOIN1", "MYCOIN", "0.8", "0.9", None),
+        ("MYCOIN1", "MYCOIN", "0.9", "0.9", None),
+        ("ETH", "MYCOIN", "0.8", "0.9", None),
+        ("MYCOIN1", "ETH", "0.8", "0.8", None),
+        ("MYCOIN1", "ETH", "0.7", "0.8", Some("0.8")),
+    ];
+    for (base, rel, price, volume, min_volume) in bob_orders.iter() {
+        let rc = block_on(mm_bob.rpc(&json! ({
+            "userpass": mm_bob.userpass,
+            "method": "setprice",
+            "base": base,
+            "rel": rel,
+            "price": price,
+            "volume": volume,
+            "min_volume": min_volume.unwrap_or("0.00777"),
+            "cancel_previous": false,
+        })))
+        .unwrap();
+        assert!(rc.0.is_success(), "!setprice: {}", rc.1);
+    }
+
+    block_on(mm_bob.wait_for_log(22., |log| {
+        log.contains("DEBUG Handling IncludedTorelaysMesh message for peer")
+    }))
+    .unwrap();
+
+    let response = block_on(best_orders_v2_by_number(&mm_alice, "MYCOIN", "buy", 1, false));
+    log!("response {response:?}");
+    let best_mycoin1_orders = response.result.orders.get("MYCOIN1").unwrap();
+    log!("Best MYCOIN1 orders when buy MYCOIN {:?}", [best_mycoin1_orders]);
+    assert_eq!(1, best_mycoin1_orders.len());
+    let expected_price: BigDecimal = "0.7".parse().unwrap();
+    assert_eq!(expected_price, best_mycoin1_orders[0].price.decimal);
+
+    let response = block_on(best_orders_v2_by_number(&mm_alice, "MYCOIN", "buy", 2, false));
+    log!("response {response:?}");
+    let best_mycoin1_orders = response.result.orders.get("MYCOIN1").unwrap();
+    log!("Best MYCOIN1 orders when buy MYCOIN {:?}", [best_mycoin1_orders]);
+    assert_eq!(2, best_mycoin1_orders.len());
+    let expected_price: BigDecimal = "0.7".parse().unwrap();
+    assert_eq!(expected_price, best_mycoin1_orders[0].price.decimal);
+    let expected_price: BigDecimal = "0.8".parse().unwrap();
+    assert_eq!(expected_price, best_mycoin1_orders[1].price.decimal);
+
+    let response = block_on(best_orders_v2_by_number(&mm_alice, "MYCOIN", "sell", 1, false));
+    log!("response {response:?}");
+    let expected_price: BigDecimal = "1.25".parse().unwrap();
+    let best_mycoin1_orders = response.result.orders.get("MYCOIN1").unwrap();
+    log!("Best MYCOIN1 orders when sell MYCOIN {:?}", [best_mycoin1_orders]);
+    assert_eq!(1, best_mycoin1_orders.len());
+    assert_eq!(expected_price, best_mycoin1_orders[0].price.decimal);
+    let best_eth_orders = response.result.orders.get("ETH").unwrap();
+    log!("Best ETH orders when sell MYCOIN {:?}", [best_eth_orders]);
+    assert_eq!(1, best_eth_orders.len());
+    assert_eq!(expected_price, best_eth_orders[0].price.decimal);
+
+    let response = block_on(best_orders_v2_by_number(&mm_alice, "ETH", "sell", 1, false));
+    log!("response {response:?}");
+    let best_mycoin_orders = response.result.orders.get("MYCOIN").unwrap();
+    log!("Best MYCOIN orders when sell ETH {:?}", [best_mycoin_orders]);
+    assert_eq!(1, best_mycoin_orders.len());
+    let expected_price: BigDecimal = "1.25".parse().unwrap();
+    assert_eq!(expected_price, best_mycoin_orders[0].price.decimal);
+
+    block_on(mm_bob.stop()).unwrap();
+    block_on(mm_alice.stop()).unwrap();
+}
+
+#[test]
+fn test_best_orders_v2_by_volume() {
+    let (mut mm_bob, mm_alice) = get_bob_alice();
+
+    // issue sell request on Bob side by setting base/rel price
+    log!("Issue bob sell requests");
+
+    let bob_orders = [
+        // (base, rel, price, volume, min_volume)
+        ("MYCOIN", "MYCOIN1", "0.9", "0.9", None),
+        ("MYCOIN", "MYCOIN1", "0.8", "0.9", None),
+        ("MYCOIN", "MYCOIN1", "0.7", "0.9", Some("0.9")),
+        ("MYCOIN", "ETH", "0.8", "0.9", None),
+        ("MYCOIN1", "MYCOIN", "0.8", "0.9", None),
+        ("MYCOIN1", "MYCOIN", "0.9", "0.9", None),
+        ("ETH", "MYCOIN", "0.8", "0.9", None),
+        ("MYCOIN1", "ETH", "0.8", "0.8", None),
+        ("MYCOIN1", "ETH", "0.7", "0.8", Some("0.8")),
+    ];
+    for (base, rel, price, volume, min_volume) in bob_orders.iter() {
+        let rc = block_on(mm_bob.rpc(&json! ({
+            "userpass": mm_bob.userpass,
+            "method": "setprice",
+            "base": base,
+            "rel": rel,
+            "price": price,
+            "volume": volume,
+            "min_volume": min_volume.unwrap_or("0.00777"),
+            "cancel_previous": false,
+        })))
+        .unwrap();
+        assert!(rc.0.is_success(), "!setprice: {}", rc.1);
+    }
+
+    block_on(mm_bob.wait_for_log(22., |log| {
+        log.contains("DEBUG Handling IncludedTorelaysMesh message for peer")
+    }))
+    .unwrap();
+
+    let response = block_on(best_orders_v2(&mm_alice, "MYCOIN", "buy", "1.7"));
+    log!("response {response:?}");
+    // MYCOIN1
+    let best_mycoin1_orders = response.result.orders.get("MYCOIN1").unwrap();
+    log!("Best MYCOIN1 orders when buy MYCOIN {:?}", [best_mycoin1_orders]);
+    let expected_price: BigDecimal = "0.7".parse().unwrap();
+    assert_eq!(expected_price, best_mycoin1_orders[0].price.decimal);
+    let expected_price: BigDecimal = "0.8".parse().unwrap();
+    assert_eq!(expected_price, best_mycoin1_orders[1].price.decimal);
+    // ETH
+    let expected_price: BigDecimal = "0.8".parse().unwrap();
+    let best_eth_orders = response.result.orders.get("ETH").unwrap();
+    log!("Best ETH orders when buy MYCOIN {:?}", [best_eth_orders]);
+    assert_eq!(expected_price, best_eth_orders[0].price.decimal);
+
+    let response = block_on(best_orders_v2(&mm_alice, "MYCOIN", "sell", "0.1"));
+    log!("response {response:?}");
+    let expected_price: BigDecimal = "1.25".parse().unwrap();
+    let best_mycoin1_orders = response.result.orders.get("MYCOIN1").unwrap();
+    log!("Best MYCOIN1 orders when sell MYCOIN {:?}", [best_mycoin1_orders]);
+    assert_eq!(expected_price, best_mycoin1_orders[0].price.decimal);
+    assert_eq!(1, best_mycoin1_orders.len());
+    let best_eth_orders = response.result.orders.get("ETH").unwrap();
+    log!("Best ETH orders when sell MYCOIN {:?}", [best_mycoin1_orders]);
+    assert_eq!(expected_price, best_eth_orders[0].price.decimal);
+
+    let response = block_on(best_orders_v2(&mm_alice, "ETH", "sell", "0.1"));
+    log!("response {response:?}");
+    let expected_price: BigDecimal = "1.25".parse().unwrap();
+    let best_mycoin1_orders = response.result.orders.get("MYCOIN1").unwrap();
+    log!("Best MYCOIN1 orders when sell ETH {:?}", [best_mycoin1_orders]);
+    assert_eq!(expected_price, best_mycoin1_orders[0].price.decimal);
+    assert_eq!("MYCOIN1", best_mycoin1_orders[0].coin);
+    assert_eq!(1, best_mycoin1_orders.len());
+
+    block_on(mm_bob.stop()).unwrap();
+    block_on(mm_alice.stop()).unwrap();
+}
+
+#[test]
+fn test_best_orders_filter_response() {
+    // alice defined MYCOIN1 as "wallet_only" in config
+    let alice_coins = json!([
+        mycoin_conf(1000),
+        {"coin":"MYCOIN1","asset":"MYCOIN1","rpcport":11608,"wallet_only": true,"txversion":4,"overwintered":1,"protocol":{"type":"UTXO"}},
+        eth_dev_conf(),
+    ]);
+
+    let bob_priv_key = random_secp256k1_secret();
+    let alice_priv_key = random_secp256k1_secret();
+
+    generate_utxo_coin_with_privkey("MYCOIN", 1000.into(), bob_priv_key);
+    generate_utxo_coin_with_privkey("MYCOIN1", 1000.into(), bob_priv_key);
+    fill_eth_erc20_with_private_key(bob_priv_key);
+
+    generate_utxo_coin_with_privkey("MYCOIN", 1000.into(), alice_priv_key);
+    generate_utxo_coin_with_privkey("MYCOIN1", 1000.into(), alice_priv_key);
+    fill_eth_erc20_with_private_key(alice_priv_key);
+
+    let coins = json!([mycoin_conf(1000), mycoin1_conf(1000), eth_dev_conf(),]);
+
+    let bob_conf = Mm2TestConf::seednode(&format!("0x{}", hex::encode(bob_priv_key)), &coins);
+    let mut mm_bob = MarketMakerIt::start(bob_conf.conf, bob_conf.rpc_password, None).unwrap();
+
+    let (_bob_dump_log, _bob_dump_dashboard) = mm_bob.mm_dump();
+    log!("Bob log path: {}", mm_bob.log_path.display());
+
+    log!("{:?}", block_on(enable_native(&mm_bob, "MYCOIN", &[], None)));
+    log!("{:?}", block_on(enable_native(&mm_bob, "MYCOIN1", &[], None)));
+    let swap_contract = swap_contract_checksum();
+    dbg!(block_on(enable_eth_coin(
+        &mm_bob,
+        "ETH",
+        &[GETH_RPC_URL],
+        &swap_contract,
+        None,
+        false
+    )));
+
+    // issue sell request on Bob side by setting base/rel price
+    log!("Issue bob sell requests");
+
+    let bob_orders = [
+        // (base, rel, price, volume, min_volume)
+        ("MYCOIN", "MYCOIN1", "0.9", "0.9", None),
+        ("MYCOIN", "MYCOIN1", "0.8", "0.9", None),
+        ("MYCOIN", "MYCOIN1", "0.7", "0.9", Some("0.9")),
+        ("MYCOIN", "ETH", "0.8", "0.9", None),
+        ("MYCOIN1", "MYCOIN", "0.8", "0.9", None),
+        ("MYCOIN1", "MYCOIN", "0.9", "0.9", None),
+        ("ETH", "MYCOIN", "0.8", "0.9", None),
+        ("MYCOIN1", "ETH", "0.8", "0.8", None),
+        ("MYCOIN1", "ETH", "0.7", "0.8", Some("0.8")),
+    ];
+    for (base, rel, price, volume, min_volume) in bob_orders.iter() {
+        let rc = block_on(mm_bob.rpc(&json! ({
+            "userpass": mm_bob.userpass,
+            "method": "setprice",
+            "base": base,
+            "rel": rel,
+            "price": price,
+            "volume": volume,
+            "min_volume": min_volume.unwrap_or("0.00777"),
+            "cancel_previous": false,
+        })))
+        .unwrap();
+        assert!(rc.0.is_success(), "!setprice: {}", rc.1);
+    }
+
+    let alice_conf = Mm2TestConf::light_node(
+        &format!("0x{}", hex::encode(alice_priv_key)),
+        &alice_coins,
+        &[&mm_bob.ip.to_string()],
+    );
+    let mm_alice = MarketMakerIt::start(alice_conf.conf, alice_conf.rpc_password, None).unwrap();
+
+    let (_alice_dump_log, _alice_dump_dashboard) = mm_alice.mm_dump();
+    log!("Alice log path: {}", mm_alice.log_path.display());
+
+    block_on(mm_bob.wait_for_log(22., |log| {
+        log.contains("DEBUG Handling IncludedTorelaysMesh message for peer")
+    }))
+    .unwrap();
+
+    let rc = block_on(mm_alice.rpc(&json! ({
+        "userpass": mm_alice.userpass,
+        "method": "best_orders",
+        "coin": "MYCOIN",
+        "action": "buy",
+        "volume": "0.1",
+    })))
+    .unwrap();
+    assert!(rc.0.is_success(), "!best_orders: {}", rc.1);
+    let response: BestOrdersResponse = serde_json::from_str(&rc.1).unwrap();
+    let empty_vec = Vec::new();
+    let best_mycoin1_orders = response.result.get("MYCOIN1").unwrap_or(&empty_vec);
+    assert_eq!(0, best_mycoin1_orders.len());
+    let best_eth_orders = response.result.get("ETH").unwrap();
+    assert_eq!(1, best_eth_orders.len());
 
     block_on(mm_bob.stop()).unwrap();
     block_on(mm_alice.stop()).unwrap();
